+++ src/MenuList.jsx
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+const MenuList = styled.ul({
+  listStyle: 'none',
+  padding: '0',
+  display: 'flex',
+  flexFlow: 'column',
+  flexWrap: 'wrap',
+  justifyContent: 'space-around',
+  alignItems: 'center',
+});
+
+export default MenuList;
