+++ src/OptionList.jsx
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const OptionList = styled.ul({
+  display: 'flex',
+  justifyContent: 'space-around',
+  alignContent: 'center',
+  margin: '30px auto',
+  padding: 0,
+  listStyle: 'none',
+});
+
+export default OptionList;
