+++ src/RegionsContainer.jsx
@@ -2,6 +2,9 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
+import OptionList from './OptionList';
+import Button from './Button';
+
 import {
   selectRegion,
   loadRestaurants,
@@ -21,10 +24,10 @@ export default function RegionsContainer() {
   }
 
   return (
-    <ul>
+    <OptionList>
       {regions.map((region) => (
         <li key={region.id}>
-          <button
+          <Button
             type="button"
             onClick={() => handleClick(region.id)}
           >
@@ -34,9 +37,9 @@ export default function RegionsContainer() {
                 {region.id === selectedRegion.id ? '(V)' : null}
               </>
             ) : null}
-          </button>
+          </Button>
         </li>
       ))}
-    </ul>
+    </OptionList>
   );
 }
